+++ src/cryojax/simulator/_volume/base_volume.py
