+++ src/cryojax/simulator/_volume/representations/base_volume.py
