+++ src/cryojax/simulator/_volume/representations/common_functions.py
