+++ src/cryojax/simulator/_volume/representations/gmm_volume.py
