+++ src/cryojax/simulator/_volume/representations/tabulated_volume.py
