+++ src/cryojax/simulator/_volume/representations/voxel_volume.py
